scx: add lib/bpf_cpumask.h for bpf_cpumask kptr lifecycle helpers
Several schedulers hand-roll the same
bpf_cpumask_create()/bpf_kptr_xchg()/bpf_cpumask_release() dance to
install a freshly allocated mask into a kptr slot. lib/percpu.h already
had such a helper (create_save_bpfmask), but it is bundled with the
arena scx_bitmap library and the per-CPU storage map, so it can't be
reused without pulling those in.

Move create_save_bpfmask() into a new lib/bpf_cpumask.h that depends
only on <scx/common.bpf.h> (the kfunc prototypes), so any scheduler can
allocate kptr cpumasks without the arena/percpu baggage. lib/percpu.h
now includes it.

This header is for the kernel "struct bpf_cpumask" kptr object and is
distinct from lib/cpumask.h, which is an arena-resident bitmap
(scx_bitmap) reimplementation.

The relocated helper drops the two debug bpf_printk()s the percpu.h copy
had; callers already propagate the -ENOMEM return.

No functional change for schedulers (other than the dropped debug
prints).
